+++ lib/modules/theme/theme_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: theme
-// Updated at: 2026-05-22 20:59:12
-// Commit: feat(theme): add async theme stream provider
+// Updated at: 2026-05-23 16:13:06
+// Commit: chore(theme): update dependencies and cleanup unused imports in theme
 
 class themeComponentState {
-  final String commitRef = '1338';
-  final String lastMessage = 'feat(theme): add async theme stream provider';
-  final DateTime timestamp = DateTime.parse('2026-05-22T20:59:12+05:00');
+  final String commitRef = '1353';
+  final String lastMessage = 'chore(theme): update dependencies and cleanup unused imports in theme';
+  final DateTime timestamp = DateTime.parse('2026-05-23T16:13:06+05:00');
 }